requirements.txt
mysqlclient==2.0.3
djangorestframework==3.12.2
django==3.1.8
python-decouple==3.4
pillow==8.1.2
django-cleanup==5.1.0
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.7.1